services/ingest/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pdfplumber==0.11.0
python-docx==1.1.0
boto3==1.34.0
psycopg2-binary==2.9.9
httpx==0.27.0
python-multipart==0.0.9
